chore(move tests): move request tests into their own folder
